+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 
-* README and LICENSE added
-* source repository and automatic packaging setup
+Tooltip alignment
+
+* Tooltip alignment is now configurable to TOP or BOTTOM (default TOP)
+* to switch tooltip alignment use /trh align <TOP|BOTTOM> (also added to the CLI help)
+* this setting will be saved account wide
+
+Reward tooltips are no longer bound to the screen to prevent overlapping with the token tooltip.